fix fluid registering
