+++ common/src/main/java/dev/latvian/kubejs/fluid/KubeJSFluidEventHandler.java
@@ -6,6 +6,9 @@ import dev.latvian.kubejs.KubeJSRegistries;
 import dev.latvian.kubejs.registry.BuilderBase;
 import dev.latvian.kubejs.registry.RegistryInfos;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.BucketItem;
+import net.minecraft.world.level.block.LiquidBlock;
+import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.material.FlowingFluid;
 import net.minecraft.world.level.material.Fluid;
 
@@ -20,7 +23,7 @@ public class KubeJSFluidEventHandler {
 	}
 
 	@ExpectPlatform
-	private static FlowingFluid buildFluid(boolean source, FluidBuilder builder) {
+	static FlowingFluid buildFluid(boolean source, FluidBuilder builder) {
 		throw new AssertionError();
 	}
 
@@ -28,8 +31,22 @@ public class KubeJSFluidEventHandler {
         for (BuilderBase<? extends Fluid> base : RegistryInfos.FLUID.objects.values()) {
             if (base instanceof FluidBuilder builder) {
                 KubeJSRegistries.fluids().register(builder.id, () -> builder.stillFluid = buildFluid(true, builder));
-                KubeJSRegistries.fluids().register(new ResourceLocation(builder.id.getNamespace(), "flowing_" + builder.id.getPath()), () -> builder.flowingFluid = buildFluid(false, builder));
+                KubeJSRegistries.fluids()
+                    .register(
+                        new ResourceLocation(builder.id.getNamespace(), "flowing_" + builder.id.getPath()),
+                        () -> builder.flowingFluid = buildFluid(false, builder)
+                    );
             }
         }
 	}
+
+    @ExpectPlatform
+    static BucketItem buildBucket(FluidBuilder builder) {
+        throw new AssertionError();
+    }
+
+    @ExpectPlatform
+    static LiquidBlock buildFluidBlock(FluidBuilder builder, BlockBehaviour.Properties properties) {
+        throw new AssertionError();
+    }
 }
