+++ forge/src/main/java/dev/latvian/kubejs/block/events/forge/KubeJSBlockEventHandlerImpl.java
@@ -1,11 +1,4 @@
 package dev.latvian.kubejs.block.events.forge;
 
-import dev.latvian.kubejs.fluid.FluidBuilder;
-import net.minecraft.world.level.block.LiquidBlock;
-import net.minecraft.world.level.block.state.BlockBehaviour;
-
 public class KubeJSBlockEventHandlerImpl {
-	public static LiquidBlock buildFluidBlock(FluidBuilder builder, BlockBehaviour.Properties properties) {
-		return new LiquidBlock(() -> builder.stillFluid, properties);
-	}
 }
